Add token to localStorage.
Make Remember Me on Login component functional.
Add remember field to the login request body based on Remember Me checkbox.
Add rememeber boolean to user schema baseJoi validaiton.
Generate a token if remember is true and add it to the response on login.
Remove token on logout.
Update token on re-login if Remember Me is checked or remove it otherwise.
Add authorization header to access and entries hooks if token is in localStorage.