Login using jwt token.
Update login controller to check if the request body has a token added from the new hasToken middleware that checks if the client is seeking authentication using the authorization header.
Have the Login react router route privately handle making the token login request if a token is found in localStorage.
Reroute back to the previous page the user was attempting to access prior to the client seeking token authentication.