Wrap Header, Body, and Footer in Document style component.
Evenly spaces the components over the entire viewport using flex.
Prevents the Footer from depending on the Body to maintain its position at the bottom of the page.
Update the Body styles component to use flexGrow instead of minHeight to occupy all the space between the header and footer in the viewport.