Add demo to Home page.
Creates a utility typeWriter function that imitates a user typing. Used to fill the value of a TextField component and then a pre-populated LLM response added to cds list of the 0th index.
Updates the information section by removing the h3 titles and replacing them with the first few words of the sentence. Rearranges each section to be side by side on md views.
Updates the margin and padding on each of the components.