Launch HTTPS enabled server in production.
In development, provided that the NODE_ENV is not production, the same http express server can be used. When NODE_ENV is production, https is enabled using the https module and loads the ssl certificates.
A new PORT environment variable is required in the backend to serve the app in development (e.g. port 3000) or to serve the app over https on the server, 8443.