Initial beta user request with email verification.
Establishes a release phase. If the release phase is beta, when a user registers they will be asked to verify their email and await a message with a special link to gain beta access through a requestBetaAccess middleware along the registration route.
The requestBetaAccess middleware creates a new user without a password preventing the user from being able to login since passport is expecting a salt value.
New fields are added to the User model for email verification that uses JWTs, verifyEmailToken, verifyEmailExpires, and emailVerified.
A new verify-email route has been added to to both the API and frontend router. The frontend simply makes a request to that access API endpoint where a verifyEmail controller sets the emailVerified User field to true. An email must be valid in order to receive the email. The client is then navigated to the Login page where they are flashed either a success or failure message.