Implement closed beta mechanism.
Creates two new routes, /beta-approval and /beta-denial. Both of which have similarly named controllers responsible for approving or denying beta user requests.
After a beta requester verifies their email address via a tokenized link, a request email is sent to a managed inbox set in the environment variables. Two magic links in the email make an API request to the new routes one that approves the request and one that denies it.
The user must respond to two emails, one to verify their email at which they will receive a magic link to change their password if they are approved for beta in the second email.
If approved the user will be asked to update their password, then they will be able to log in, otherwise, they will not and any attempts to login or register with the same email will be rejected.
Users have a week to respond to either message else the token will expire essentially denying access to beta.
Emails sent out have been updated to be from the SMTP_NAME in the environment variables.
Fixes a bug that causes the app to crash if no journal exists on login after being approved for beta.
Updates User model with required token variables, improves naming, and includes a new betaAccess token verifying beta access.