Give existing users beta access.
For existing users before beta was closed to the public, along the login or token-login paths, if the release phase is beta, the user will be granted beta access using the routes associated controller.
Rearranges the passport logIn body in tokenLogin to return next error if errored before setting flash messages.