Protect against backdoor attack,
Prevents sus user from creating an account required for beta use then using the forgot password mechanism to gain access.
If this activity is suspected, an email is sent to the admin with a link to deny or approve the user beta access.
Updates related pre-approved beta access events to status code 403.