Focus recent thought on click.
A user may now click on a recent thought to focus it in the analysis view. This functions the same as if hitting the focus button.
Additionally, when a recent thought is highlighted and in focus, the focus button is removed. This helps signify that the function of the button has been utilized.