Reduce single Entry API calls to one.
Updates the resources returned for the API endpoint to view a single entry returning all associated documents.
Makes only one API call to the updated endpoint in the Analysis component and updates how the retrieved data is inserted into the components state variables.
Updates how the Thoughts component inserts the data retrieved from the same endpoint as a result of the changes.