Update entries model to associate analysis and chat in schema.
Updating the entries model allows the use of populate which reduces the number of queries made on the database.
This update required changes to each POST API along the entries route to save as a reference the associated documents in each entry document.
For every API call on the frontend, data insertion into the components state had to be updated to reflect the new data structure returned from using populate on entries.
Additionally, typeWrittenId is set empty in Thoughts to prevent typewriter effect from occuring after en entry has been edited.