Update the analysis seed to enforce analysis topics.
Enforces json enabled gpts to return all required analysis topics, a reframing, a label and description of the cognitive distortion, an assessment of the impact having that cognitive distortion may cause, or a positive affiramtion if the thinking is healthy.
Removes analyis_content from the instruction format and replaces it with reframed_thought, distortion_analysis, and impact_assessment used to form the analysis_content stored in the db on return.
As a result of forming the analysis_content after retrieving the gpt content, json is being parsed within the entryAnalysis model methods previously parsed in the entry controllers.