Add default title to journal.
Fixes a bug where the initial title of a new journal does not populate. This was likely due it not being required and the JOI validation having conflicting rules.
JOI validation removes allow and empty rules for empty strings and instead depends on the default value to populate the title if it does happen to be empty.