Initial reports view.
Creates a Dashboard component to display the reports.
Creates a DashboardRoute to load the Dashboard components.
Adds a dashboard nav item and navigation to /reports.
Adds a new custom hook to fetch data at reports api.
Adds an initial dashboard page with a button to view a weekly report.
Centers and adds margin to nav items in unstacked view.