[150-containerize-the-cdj] Deploy replica set MDB
MongoDB is deployed as a single instance replica set, with another
service, mongosetup, initializing the DB container. Hostnames in the
README.Docker.md and compose file were updated to fit the feature.

Delete calls were not working properly, possibly due to a configuration
requiring writes to have replication. Rather than changing the config,
deploying a replica set in the local environment was chosen to mirror
the production environment.