[150-containerize-the-cdj] Init with bash script
Removed container that initiated replica set and replaced with a bash
script that the user runs to initiate replica set and seed db after
spinning up containers. It was decided that the automation gained from
the init container wasn't worth having an extra container spin up
every time.

Seed.js is run within backend container, which allows for db seeding
without needing nodejs installed on docker host. This also resolves
an issue with the hostnames now that mongodb is run as replica set.