Resolve issue with module paths in jest (#139)
Jest seems to have an issue resolving paths for typescript modules
getting imported with a .js path after transpilation. This seems
to be a common issue for many users, and a common way of getting
around it is by setting the moduleNameMapper key in the jest config
to drop the file extension for .js imports.