Remove seed.js from Jest global scripts (#139)
Using seed.js in jest globalSetup and globalTeardown scripts was causing
issues with imports because they seem to run before code transformation.
This means that any files imported in seed.js that have .ts extension
will break the script when running unit tests.

This commit removes seeding the db from the globalSetup and
globalTeardown scripts, and leaves setting up the memory mongodb
in them. It also rewrites index.test.ts to seed the db before
running tests, as an example for how tests that use the db should
set up and clean up.