Remove getPort.
Removed get port and moved port assignment into the startServer function.
Removed the Server import and assigned server to a returned instance of the Server obj
ect created in startServer.
Throw an error if the server is listening on a non-IP socket.
Update the condition block for NODE_ENV === 'production' to reflect changes made in no
n-production environments.
Reorganized file: moved all function definitions to the bottom of the file.
Updated server listening messages for more clarity.