Let Express set port for test environment. (#160)
* Use arguments to run tests on separate port with yargs.

Removes MEMORY environment variable.
Uses yargs to pass arguments to the application on launch.
 - Adds yargs to dependecies list.
 - Passes port=3001 to test instances of the app.
Resolves lint error in index.test.ts regarding imports out of order.

* Remove yargs dependency.

* Refactor server startup to dynamically assign port.

Adds new getPort and isPortInUse utility functions to dynamically assign a port in a test environment.
Updates index.ts to use the getPort function in a new async startServer function that wraps the pre-existing code.
Updates the index.ts test from a hardcoded port to matching the console log for a port between 1 to 5 digits.

* Add getPort unit tests.

* Let express assign port in test env (#139)

Rewrite getPort to allow express to assign open port in test environment
by returning 0.

* Remove getPort.

Removed get port and moved port assignment into the startServer function.
Removed the Server import and assigned server to a returned instance of the Server obj
ect created in startServer.
Throw an error if the server is listening on a non-IP socket.
Update the condition block for NODE_ENV === 'production' to reflect changes made in no
n-production environments.
Reorganized file: moved all function definitions to the bottom of the file.
Updated server listening messages for more clarity.

---------

Co-authored-by: David Lu <ludavid33@gmail.com>