refactor: Start adding tests for entry controller
Added tests for getting and creating entries. getAnalysisContent
in EntryAnalysis was breaking test suite for some reason. It appeared
that using Config in the method caused an issue with the model
compilation within Jest runtime, but I'm not sure why. I pulled it out
from the file and put it in the entry services to fix it.

In-progress, so may be broken.