fix: resolve build issue with Express.User type
The Passport module needs to be aware of our implementation of users
as defined in the User model. This was partially implemented by extending
Express.User to have an id property in backend/src/types/express/index.d.ts.
The reason for adding just the id was that it was the only property
getting used in the access controller. However, this caused issues in
app.ts because it expects Express.User to match our User model type.

This commit extends Express.User with UserType, which lets Passport
access our User definition. It also cleans up type annotations affected
by this change.