refactor: push error handling in createEntry into service
Having createEntry handle errors from the analysis completion was
making it difficult to disentangle making the Entry and EntryAnalysis
from terminating the express request. There were a few ways of resolving
this, but I decided to push the error handling into the entry service
module, and have it return an error message. This way I could simplify
the flow in the controller.