refactor: create service layer for read operations (#187)
* refactor: Start adding tests for entry controller

Added tests for getting and creating entries. getAnalysisContent
in EntryAnalysis was breaking test suite for some reason. It appeared
that using Config in the method caused an issue with the model
compilation within Jest runtime, but I'm not sure why. I pulled it out
from the file and put it in the entry services to fix it.

In-progress, so may be broken.

* refactor: add documentation for entry services

Added JSDocs for functions in entry services and broke up
createEntryAnalysis into 2 functions.

* refactor: create getPopulatedEntry function

* refactor: write functions for entry reads

Refactored entry controllers to move read calls into functions in
entry services

* revert: createEntry controller to ts migration branch

To limit the scope of this branch, I reverted changes I made to the
createEntry function in the entry contorllers to match the
139-ts-migration branch

* refactor: add more comments entry service

* Apply suggestions from code review

Co-authored-by: Ryan James Meneses <58452495+hiyaryan@users.noreply.github.com>

---------

Co-authored-by: Ryan James Meneses <58452495+hiyaryan@users.noreply.github.com>